% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis7-package.R
\docType{package}
\name{basis7-package}
\alias{basis7}
\alias{basis7-package}
\title{basis7: An S7 Framework for Basis Expansions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements basis expansions using the S7 object-oriented system. Bases are objects that can be evaluated, differentiated to any order, integrated, and asked for the Gram matrices of their derivatives, which is what a smoothness penalty integrates. Exact formulas are used wherever they exist, with numerical fallbacks that make a user-defined basis work from its evaluation alone, and a diagnostic that verifies every component.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statmodels7.github.io/basis7/}
  \item \url{https://github.com/statmodels7/basis7}
  \item Report bugs at \url{https://github.com/statmodels7/basis7/issues}
}

}
\author{
\strong{Maintainer}: Giovanni Tinervia \email{giovannitinervia9@gmail.com}

}
\keyword{internal}
