% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{basis_colnames.basis}
\alias{basis_colnames.basis}
\title{Default Column Names}
\arguments{
\item{basis}{An object inheriting from class \code{basis}.}

\item{...}{Unused.}
}
\value{
A character vector of length \code{basis@dimension}.
}
\description{
Numbers the basis functions after the family name, as in \code{bs1},
\code{bs2}, and so on.
}
\keyword{internal}
