% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_basis.R
\name{basis_colnames.PolyBasis}
\alias{basis_colnames.PolyBasis}
\title{Column Names of a Legendre Basis}
\arguments{
\item{basis}{A \code{\link[=PolyBasis]{PolyBasis()}} object.}

\item{...}{Unused.}
}
\value{
A character vector of length \code{basis@dimension}.
}
\description{
The polynomials numbered by their degree, from zero.
}
\keyword{internal}
