% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_basis.R
\name{print_basis_checks}
\alias{print_basis_checks}
\title{Print the Outcome of check_basis}
\usage{
print_basis_checks(basis, res, num)
}
\arguments{
\item{basis}{An object inheriting from class \code{basis}.}

\item{res}{The logical vector of results.}

\item{num}{The logical vector of numerical fallbacks.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Print the Outcome of check_basis
}
\keyword{internal}
